bi-gcp-application/Data_Extraction/requirements.txt
pyodbc
pandas
google-cloud-storage
pyarrow
google.cloud.logging
multiprocess
mysql-connector-python
google-cloud-secret-manager